+++ lib/src/main/java/com/diffplug/spotless/Formatter.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 DiffPlug
+ * Copyright 2016-2023 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -24,16 +24,9 @@ import java.io.ObjectOutputStream;
 import java.io.ObjectStreamException;
 import java.io.Serializable;
 import java.nio.charset.Charset;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.nio.file.StandardOpenOption;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
-
-import javax.annotation.Nullable;
 
 /** Formatter which performs the full formatting. */
 public final class Formatter implements Serializable, AutoCloseable {
@@ -41,25 +34,19 @@ public final class Formatter implements Serializable, AutoCloseable {
 
 	private LineEnding.Policy lineEndingsPolicy;
 	private Charset encoding;
-	private Path rootDir;
 	private List<FormatterStep> steps;
-	private FormatExceptionPolicy exceptionPolicy;
 
-	private Formatter(LineEnding.Policy lineEndingsPolicy, Charset encoding, Path rootDirectory, List<FormatterStep> steps, FormatExceptionPolicy exceptionPolicy) {
+	private Formatter(LineEnding.Policy lineEndingsPolicy, Charset encoding, List<FormatterStep> steps) {
 		this.lineEndingsPolicy = Objects.requireNonNull(lineEndingsPolicy, "lineEndingsPolicy");
 		this.encoding = Objects.requireNonNull(encoding, "encoding");
-		this.rootDir = Objects.requireNonNull(rootDirectory, "rootDir");
 		this.steps = requireElementsNonNull(new ArrayList<>(steps));
-		this.exceptionPolicy = Objects.requireNonNull(exceptionPolicy, "exceptionPolicy");
 	}
 
 	// override serialize output
 	private void writeObject(ObjectOutputStream out) throws IOException {
 		out.writeObject(lineEndingsPolicy);
 		out.writeObject(encoding.name());
-		out.writeObject(rootDir.toString());
 		out.writeObject(steps);
-		out.writeObject(exceptionPolicy);
 	}
 
 	// override serialize input
@@ -67,9 +54,7 @@ public final class Formatter implements Serializable, AutoCloseable {
 	private void readObject(ObjectInputStream in) throws IOException, ClassNotFoundException {
 		lineEndingsPolicy = (LineEnding.Policy) in.readObject();
 		encoding = Charset.forName((String) in.readObject());
-		rootDir = Paths.get((String) in.readObject());
 		steps = (List<FormatterStep>) in.readObject();
-		exceptionPolicy = (FormatExceptionPolicy) in.readObject();
 	}
 
 	// override serialize input
@@ -86,16 +71,8 @@ public final class Formatter implements Serializable, AutoCloseable {
 		return encoding;
 	}
 
-	public Path getRootDir() {
-		return rootDir;
-	}
-
 	public List<FormatterStep> getSteps() {
 		return steps;
-	}
-
-	public FormatExceptionPolicy getExceptionPolicy() {
-		return exceptionPolicy;
 	}
 
 	public static Formatter.Builder builder() {
@@ -106,9 +83,7 @@ public final class Formatter implements Serializable, AutoCloseable {
 		// required parameters
 		private LineEnding.Policy lineEndingsPolicy;
 		private Charset encoding;
-		private Path rootDir;
 		private List<FormatterStep> steps;
-		private FormatExceptionPolicy exceptionPolicy;
 
 		private Builder() {}
 
@@ -122,84 +97,13 @@ public final class Formatter implements Serializable, AutoCloseable {
 			return this;
 		}
 
-		public Builder rootDir(Path rootDir) {
-			this.rootDir = rootDir;
-			return this;
-		}
-
 		public Builder steps(List<FormatterStep> steps) {
 			this.steps = steps;
 			return this;
 		}
 
-		public Builder exceptionPolicy(FormatExceptionPolicy exceptionPolicy) {
-			this.exceptionPolicy = exceptionPolicy;
-			return this;
-		}
-
 		public Formatter build() {
-			return new Formatter(lineEndingsPolicy, encoding, rootDir, steps,
-					exceptionPolicy == null ? FormatExceptionPolicy.failOnlyOnError() : exceptionPolicy);
-		}
-	}
-
-	/** Returns true iff the given file's formatting is up-to-date. */
-	public boolean isClean(File file) throws IOException {
-		Objects.requireNonNull(file);
-
-		String raw = new String(Files.readAllBytes(file.toPath()), encoding);
-		String unix = LineEnding.toUnix(raw);
-
-		// check the newlines (we can find these problems without even running the steps)
-		int totalNewLines = (int) unix.codePoints().filter(val -> val == '\n').count();
-		int windowsNewLines = raw.length() - unix.length();
-		if (lineEndingsPolicy.isUnix(file)) {
-			if (windowsNewLines != 0) {
-				return false;
-			}
-		} else {
-			if (windowsNewLines != totalNewLines) {
-				return false;
-			}
-		}
-
-		// check the other formats
-		String formatted = compute(unix, file);
-
-		// return true iff the formatted string equals the unix one
-		return formatted.equals(unix);
-	}
-
-	/** Applies formatting to the given file. */
-	public void applyTo(File file) throws IOException {
-		applyToAndReturnResultIfDirty(file);
-	}
-
-	/**
-	 * Applies formatting to the given file.
-	 *
-	 * Returns null if the file was already clean, or the
-	 * formatted result with unix newlines if it was not.
-	 */
-	public @Nullable String applyToAndReturnResultIfDirty(File file) throws IOException {
-		Objects.requireNonNull(file);
-
-		byte[] rawBytes = Files.readAllBytes(file.toPath());
-		String raw = new String(rawBytes, encoding);
-		String rawUnix = LineEnding.toUnix(raw);
-
-		// enforce the format
-		String formattedUnix = compute(rawUnix, file);
-		// enforce the line endings
-		String formatted = computeLineEndings(formattedUnix, file);
-
-		// write out the file iff it has changed
-		byte[] formattedBytes = formatted.getBytes(encoding);
-		if (!Arrays.equals(rawBytes, formattedBytes)) {
-			Files.write(file.toPath(), formattedBytes, StandardOpenOption.TRUNCATE_EXISTING);
-			return formattedUnix;
-		} else {
-			return null;
+			return new Formatter(lineEndingsPolicy, encoding, steps);
 		}
 	}
 
@@ -237,8 +141,8 @@ public final class Formatter implements Serializable, AutoCloseable {
 					unix = LineEnding.toUnix(formatted);
 				}
 			} catch (Throwable e) {
-				String relativePath = rootDir.relativize(file.toPath()).toString();
-				exceptionPolicy.handleError(e, step, relativePath);
+				// TODO: this is bad, but it won't matter when add support for linting
+				throw new RuntimeException(e);
 			}
 		}
 		return unix;
@@ -250,9 +154,7 @@ public final class Formatter implements Serializable, AutoCloseable {
 		int result = 1;
 		result = prime * result + encoding.hashCode();
 		result = prime * result + lineEndingsPolicy.hashCode();
-		result = prime * result + rootDir.hashCode();
 		result = prime * result + steps.hashCode();
-		result = prime * result + exceptionPolicy.hashCode();
 		return result;
 	}
 
@@ -270,9 +172,7 @@ public final class Formatter implements Serializable, AutoCloseable {
 		Formatter other = (Formatter) obj;
 		return encoding.equals(other.encoding) &&
 				lineEndingsPolicy.equals(other.lineEndingsPolicy) &&
-				rootDir.equals(other.rootDir) &&
-				steps.equals(other.steps) &&
-				exceptionPolicy.equals(other.exceptionPolicy);
+				steps.equals(other.steps);
 	}
 
 	@SuppressWarnings("rawtypes")