+++ testlib/src/test/java/com/diffplug/spotless/kotlin/DiktatStepTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021-2022 DiffPlug
+ * Copyright 2021-2023 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -25,27 +25,19 @@ import org.junit.jupiter.api.Test;
 import com.diffplug.spotless.FileSignature;
 import com.diffplug.spotless.FormatterStep;
 import com.diffplug.spotless.ResourceHarness;
-import com.diffplug.spotless.StepHarness;
+import com.diffplug.spotless.StepHarnessWithFile;
 import com.diffplug.spotless.TestProvisioner;
 
 class DiktatStepTest extends ResourceHarness {
 
 	@Test
-	void behavior() throws Exception {
+	void behavior() {
 		FormatterStep step = DiktatStep.create(TestProvisioner.mavenCentral());
-		StepHarness.forStep(step)
-				.testResourceException("kotlin/diktat/Unsolvable.kt", assertion -> {
-					assertion.isInstanceOf(AssertionError.class);
-					assertion.hasMessage("There are 4 unfixed errors:" +
-							System.lineSeparator() + "Error on line: 1, column: 1 cannot be fixed automatically" +
-							System.lineSeparator() + "[FILE_NAME_INCORRECT] file name is incorrect - it should end with .kt extension and be in PascalCase: testlib" +
-							System.lineSeparator() + "Error on line: 1, column: 1 cannot be fixed automatically" +
-							System.lineSeparator() + "[FILE_NAME_MATCH_CLASS] file name is incorrect - it should match with the class described in it if there is the only one class declared: testlib vs Unsolvable" +
-							System.lineSeparator() + "Error on line: 1, column: 1 cannot be fixed automatically" +
-							System.lineSeparator() + "[DEBUG_PRINT] use a dedicated logging library: found println()" +
-							System.lineSeparator() + "Error on line: 13, column: 9 cannot be fixed automatically" +
-							System.lineSeparator() + "[DEBUG_PRINT] use a dedicated logging library: found println()");
-				});
+		StepHarnessWithFile.forStep(this, step).testResourceExceptionMsg("kotlin/diktat/Unsolvable.kt").isEqualTo("There are 2 unfixed errors:" +
+				System.lineSeparator() + "Error on line: 1, column: 1 cannot be fixed automatically" +
+				System.lineSeparator() + "[DEBUG_PRINT] use a dedicated logging library: found println()" +
+				System.lineSeparator() + "Error on line: 13, column: 9 cannot be fixed automatically" +
+				System.lineSeparator() + "[DEBUG_PRINT] use a dedicated logging library: found println()");
 	}
 
 	@Test
@@ -55,19 +47,11 @@ class DiktatStepTest extends ResourceHarness {
 		FileSignature config = signAsList(conf);
 
 		FormatterStep step = DiktatStep.create("1.2.1", TestProvisioner.mavenCentral(), config);
-		StepHarness.forStep(step)
-				.testResourceException("kotlin/diktat/Unsolvable.kt", assertion -> {
-					assertion.isInstanceOf(AssertionError.class);
-					assertion.hasMessage("There are 4 unfixed errors:" +
-							System.lineSeparator() + "Error on line: 1, column: 1 cannot be fixed automatically" +
-							System.lineSeparator() + "[FILE_NAME_INCORRECT] file name is incorrect - it should end with .kt extension and be in PascalCase: testlib" +
-							System.lineSeparator() + "Error on line: 1, column: 1 cannot be fixed automatically" +
-							System.lineSeparator() + "[FILE_NAME_MATCH_CLASS] file name is incorrect - it should match with the class described in it if there is the only one class declared: testlib vs Unsolvable" +
-							System.lineSeparator() + "Error on line: 1, column: 1 cannot be fixed automatically" +
-							System.lineSeparator() + "[DEBUG_PRINT] use a dedicated logging library: found println()" +
-							System.lineSeparator() + "Error on line: 13, column: 9 cannot be fixed automatically" +
-							System.lineSeparator() + "[DEBUG_PRINT] use a dedicated logging library: found println()");
-				});
+		StepHarnessWithFile.forStep(this, step).testResourceExceptionMsg("kotlin/diktat/Unsolvable.kt").isEqualTo("There are 2 unfixed errors:" +
+				System.lineSeparator() + "Error on line: 1, column: 1 cannot be fixed automatically" +
+				System.lineSeparator() + "[DEBUG_PRINT] use a dedicated logging library: found println()" +
+				System.lineSeparator() + "Error on line: 13, column: 9 cannot be fixed automatically" +
+				System.lineSeparator() + "[DEBUG_PRINT] use a dedicated logging library: found println()");
 	}
 
 	@Test