+++ CHANGES.md
@@ -9,6 +9,14 @@ This document is intended for Spotless developers.
 
 We adhere to the [keepachangelog](https://keepachangelog.com/en/1.0.0/) format (starting after version `1.27.0`).
 
+## [TBD lint release]
+* **BREAKING** Removed all deprecated methods.
+* **BREAKING** Removed `isClean`, `applyTo`, and `applyToAndReturnResultIfDirty` from `Formatter` because users should instead use new class `DirtyState.of`
+* **BREAKING** Removed `FormatterStep.Strict` because it was an unnecessary and unused implementation detail
+* **BREAKING** Removed `rootDir` and `exceptionPolicy` parameters from `Formatter`
+* **BREAKING** Renamed `PipeStepPair` to `FenceStep` and changed its approach to be lint-friendly
+* **BREAKING** Modified `StepHarness` and `StepHarnessWithFile` to stop asserting on exceptions (to prepare for Lint)
+
 ## [Unreleased]
 ### Changes
 * Bump the dev version of Gradle from `7.5.1` to `7.6` ([#1409](https://github.com/diffplug/spotless/pull/1409))